+++ admin/Makefile
@@ -1,4 +1,9 @@
 
 .PHONY: build-development
 build-development:
-	NODE_ENV=development pnpm build
+	pnpm build
+
+.PHONY: deploy-development
+deploy-development:
+	aws s3 sync ./out/  s3://admin.reaction-development.swiswiswift.com --exact-timestamps --delete --profile reaction-development  
+	# aws cloudfront create-invalidation --distribution-id E7K0W7JYJMRK --paths "/resource/*" --profile chemist
